adapt docson for npm
* use browserify in example and test
* remove most files in lib/ : now use npm packages instead
* remove some files that cannot really be exported and probably belong in other packages (swagger, widget)
* simplify the example
