+++ .gitignore
@@ -1,3 +1,4 @@
 .idea
 docson.iml
 /nbproject
+bundle.js
