+++ example/example.js
@@ -1,0 +1,23 @@
+var docson=require("docson");
+
+var element=document.body.appendChild(document.createElement("div"));
+var schema={
+  "title": "Example Schema",
+  "type": "object",
+  "properties": {
+    "firstName": {
+      "type": "string"
+    },
+    "lastName": {
+      "type": "string"
+    },
+    "age": {
+      "description": "Age in years",
+      "type": "integer",
+      "minimum": 0
+    }
+  },
+  "required": ["firstName", "lastName"]
+};
+
+docson.doc(element,schema);
