--- examples/patternProperties.json
@@ -1,52 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "Standing instance",
-    "type": "object",
-    "required": ["positions"],
-    "properties": {
-        "positions": {
-            "type": "object",
-            "description": "Docson is able to visualize it",
-            "properties": {
-                "id": {
-                    "description": "The unique identifier for a product",
-                    "type": "number"
-                },
-                "name": {
-                    "type": "string"
-                }
-            },
-            "patternProperties": {
-                "a*": {"type": "integer"},
-                "aaa*": {"type": "integer", "maximum": 20},
-                "[0-9]{2,}": { "type": "boolean" },
-                "X_": { "type": "string" },
-                "[0-9]+": {
-                    "type": "array",
-                    "minItems": 1,
-                    "items": {
-                        "type": "object",
-                        "required": ["key"],
-                        "additionalProperties": false,
-                        "properties": {
-                            "key": {
-                                "type": "integer"
-                            }
-                        }
-                    }
-                }
-            },
-            "additionalProperties": false
-        }
-    },
-    "patternProperties": {
-        "a*": {"oneOf": [
-            { "type": "integer", "maximum": 20},
-            { "type": "boolean" },
-            { "type": "string", "maxLength": 5}
-        ]},
-        "aaa*": {"type": "integer", "maximum": 20},
-        "[0-9]{2,}": { "type": "boolean" },
-        "X_": { "type": "string" }
-    }
-}
