--- examples/properties.json
@@ -1,14 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "Standing instance",
-    "type": "object",
-    "required": ["positions"],
-    "properties": {
-        "positions": {
-            "anyOf": [
-                { "type": "integer", "maximum": 20, "description": "this is a desc"},
-                { "type": "boolean", "description": "this is a desc" }]
-        }
-    },
-    "additionalProperties": false
-}
