+++ package.json
@@ -1,0 +1,45 @@
+{
+  "name": "docson",
+  "version": "1.0.0",
+  "description": "Documentation for your JSON types.",
+  "main": "docson.js",
+  "directories": {
+    "example": "examples",
+    "test": "tests"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "prepublish": "require-self"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/lbovet/docson.git"
+  },
+  "keywords": [
+    "json",
+    "schema",
+    "documentation"
+  ],
+  "browserify": {
+    "transform": [
+      "brfs"
+    ]
+  },
+  "author": "Laurent Bovet",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/lbovet/docson/issues"
+  },
+  "homepage": "https://github.com/lbovet/docson#readme",
+  "devDependencies": {
+    "brfs": "^1.4.3",
+    "require-self": "^0.1.0"
+  },
+  "dependencies": {
+    "handlebars": "^4.0.5",
+    "jquery": "^2.2.2",
+    "jsonpointer.js": "^0.4.0",
+    "marked": "^0.3.5",
+    "traverse": "^0.6.6"
+  }
+}
